+++ examples/ae_mnist.py
