+++ examples/variational_ae_mnist.py
